# internal/goa/service/resume/gen/http/hy_usertech/server/server.go
// Code generated by goa v3.7.13, DO NOT EDIT.
//
// hy_usertech HTTP server
//
// Command:
// $ goa gen resume/design

package server

import (
	"context"
	"net/http"
	"regexp"
	hyusertech "resume/gen/hy_usertech"

	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
	"goa.design/plugins/v3/cors"
)

// Server lists the hy_usertech service endpoint HTTP handlers.
type Server struct {
	Mounts             []*MountPoint
	GetUserLikeTech    http.Handler
	GetUserDisLikeTech http.Handler
	CORS               http.Handler
}

// ErrorNamer is an interface implemented by generated error structs that
// exposes the name of the error as defined in the design.
type ErrorNamer interface {
	ErrorName() string
}

// MountPoint holds information about the mounted endpoints.
type MountPoint struct {
	// Method is the name of the service method served by the mounted HTTP handler.
	Method string
	// Verb is the HTTP method used to match requests to the mounted handler.
	Verb string
	// Pattern is the HTTP request path pattern used to match requests to the
	// mounted handler.
	Pattern string
}

// New instantiates HTTP handlers for all the hy_usertech service endpoints
// using the provided encoder and decoder. The handlers are mounted on the
// given mux using the HTTP verb and path defined in the design. errhandler is
// called whenever a response fails to be encoded. formatter is used to format
// errors returned by the service methods prior to encoding. Both errhandler
// and formatter are optional and can be nil.
func New(
	e *hyusertech.Endpoints,
	mux goahttp.Muxer,
	decoder func(*http.Request) goahttp.Decoder,
	encoder func(context.Context, http.ResponseWriter) goahttp.Encoder,
	errhandler func(context.Context, http.ResponseWriter, error),
	formatter func(err error) goahttp.Statuser,
) *Server {
	return &Server{
		Mounts: []*MountPoint{
			{"GetUserLikeTech", "GET", "/user/{userID}/liketech"},
			{"GetUserDisLikeTech", "GET", "/user/{userID}/disliketech"},
			{"CORS", "OPTIONS", "/user/{userID}/liketech"},
			{"CORS", "OPTIONS", "/user/{userID}/disliketech"},
		},
		GetUserLikeTech:    NewGetUserLikeTechHandler(e.GetUserLikeTech, mux, decoder, encoder, errhandler, formatter),
		GetUserDisLikeTech: NewGetUserDisLikeTechHandler(e.GetUserDisLikeTech, mux, decoder, encoder, errhandler, formatter),
		CORS:               NewCORSHandler(),
	}
}

// Service returns the name of the service served.
func (s *Server) Service() string { return "hy_usertech" }

// Use wraps the server handlers with the given middleware.
func (s *Server) Use(m func(http.Handler) http.Handler) {
	s.GetUserLikeTech = m(s.GetUserLikeTech)
	s.GetUserDisLikeTech = m(s.GetUserDisLikeTech)
	s.CORS = m(s.CORS)
}

// Mount configures the mux to serve the hy_usertech endpoints.
func Mount(mux goahttp.Muxer, h *Server) {
	MountGetUserLikeTechHandler(mux, h.GetUserLikeTech)
	MountGetUserDisLikeTechHandler(mux, h.GetUserDisLikeTech)
	MountCORSHandler(mux, h.CORS)
}

// Mount configures the mux to serve the hy_usertech endpoints.
func (s *Server) Mount(mux goahttp.Muxer) {
	Mount(mux, s)
}

// MountGetUserLikeTechHandler configures the mux to serve the "hy_usertech"
// service "getUserLikeTech" endpoint.
func MountGetUserLikeTechHandler(mux goahttp.Muxer, h http.Handler) {
	f, ok := HandleHyUsertechOrigin(h).(http.HandlerFunc)
	if !ok {
		f = func(w http.ResponseWriter, r *http.Request) {
			h.ServeHTTP(w, r)
		}
	}
	mux.Handle("GET", "/user/{userID}/liketech", f)
}

// NewGetUserLikeTechHandler creates a HTTP handler which loads the HTTP
// request and calls the "hy_usertech" service "getUserLikeTech" endpoint.
func NewGetUserLikeTechHandler(
	endpoint goa.Endpoint,
	mux goahttp.Muxer,
	decoder func(*http.Request) goahttp.Decoder,
	encoder func(context.Context, http.ResponseWriter) goahttp.Encoder,
	errhandler func(context.Context, http.ResponseWriter, error),
	formatter func(err error) goahttp.Statuser,
) http.Handler {
	var (
		decodeRequest  = DecodeGetUserLikeTechRequest(mux, decoder)
		encodeResponse = EncodeGetUserLikeTechResponse(encoder)
		encodeError    = goahttp.ErrorEncoder(encoder, formatter)
	)
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := context.WithValue(r.Context(), goahttp.AcceptTypeKey, r.Header.Get("Accept"))
		ctx = context.WithValue(ctx, goa.MethodKey, "getUserLikeTech")
		ctx = context.WithValue(ctx, goa.ServiceKey, "hy_usertech")
		payload, err := decodeRequest(r)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		res, err := endpoint(ctx, payload)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		if err := encodeResponse(ctx, w, res); err != nil {
			errhandler(ctx, w, err)
		}
	})
}

// MountGetUserDisLikeTechHandler configures the mux to serve the "hy_usertech"
// service "getUserDisLikeTech" endpoint.
func MountGetUserDisLikeTechHandler(mux goahttp.Muxer, h http.Handler) {
	f, ok := HandleHyUsertechOrigin(h).(http.HandlerFunc)
	if !ok {
		f = func(w http.ResponseWriter, r *http.Request) {
			h.ServeHTTP(w, r)
		}
	}
	mux.Handle("GET", "/user/{userID}/disliketech", f)
}

// NewGetUserDisLikeTechHandler creates a HTTP handler which loads the HTTP
// request and calls the "hy_usertech" service "getUserDisLikeTech" endpoint.
func NewGetUserDisLikeTechHandler(
	endpoint goa.Endpoint,
	mux goahttp.Muxer,
	decoder func(*http.Request) goahttp.Decoder,
	encoder func(context.Context, http.ResponseWriter) goahttp.Encoder,
	errhandler func(context.Context, http.ResponseWriter, error),
	formatter func(err error) goahttp.Statuser,
) http.Handler {
	var (
		decodeRequest  = DecodeGetUserDisLikeTechRequest(mux, decoder)
		encodeResponse = EncodeGetUserDisLikeTechResponse(encoder)
		encodeError    = goahttp.ErrorEncoder(encoder, formatter)
	)
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := context.WithValue(r.Context(), goahttp.AcceptTypeKey, r.Header.Get("Accept"))
		ctx = context.WithValue(ctx, goa.MethodKey, "getUserDisLikeTech")
		ctx = context.WithValue(ctx, goa.ServiceKey, "hy_usertech")
		payload, err := decodeRequest(r)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		res, err := endpoint(ctx, payload)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		if err := encodeResponse(ctx, w, res); err != nil {
			errhandler(ctx, w, err)
		}
	})
}

// MountCORSHandler configures the mux to serve the CORS endpoints for the
// service hy_usertech.
func MountCORSHandler(mux goahttp.Muxer, h http.Handler) {
	h = HandleHyUsertechOrigin(h)
	mux.Handle("OPTIONS", "/user/{userID}/liketech", h.ServeHTTP)
	mux.Handle("OPTIONS", "/user/{userID}/disliketech", h.ServeHTTP)
}

// NewCORSHandler creates a HTTP handler which returns a simple 200 response.
func NewCORSHandler() http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		w.WriteHeader(200)
	})
}

// HandleHyUsertechOrigin applies the CORS response headers corresponding to
// the origin for the service hy_usertech.
func HandleHyUsertechOrigin(h http.Handler) http.Handler {
	spec0 := regexp.MustCompile("swagger.goa.design")
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		origin := r.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			h.ServeHTTP(w, r)
			return
		}
		if cors.MatchOriginRegexp(origin, spec0) {
			w.Header().Set("Access-Control-Allow-Origin", origin)
			w.Header().Set("Vary", "Origin")
			w.Header().Set("Access-Control-Max-Age", "600")
			w.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := r.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE")
				w.Header().Set("Access-Control-Allow-Headers", "*")
			}
			h.ServeHTTP(w, r)
			return
		}
		h.ServeHTTP(w, r)
		return
	})
}