# internal/goa/service/resume/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: resume API
    description: go-goa API with goa framework
    contact:
        name: hiromaily
        url: https://hiromaily.github.io/
        email: hiromaily@gmail.com
    license:
        name: MIT
        url: https://github.com/goadesign/goa/blob/master/LICENSE
    version: "1.0"
servers:
    - url: http://localhost:8080/api
paths:
...
    /user/{userID}/disliketech:
        get:
            tags:
                - hy_usertech
            summary: getUserDisLikeTech hy_usertech
            description: get user's dislike techs
            operationId: hy_usertech#getUserDisLikeTech
            parameters:
                - name: userID
                  in: path
                  description: User ID
                  required: true
                  schema:
                    type: integer
                    description: User ID
                    example: 440487564314040275
                    minimum: 1
                  example: 6273665427536077510
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                description: 'Response body may alternatively be #/components/schemas/UsertechResponseTechCollection'
                                $ref: '#/components/schemas/UsertechCollection'
                            example:
                                - id: 10
                                  tech_name: Golang
                                - id: 10
                                  tech_name: Golang
            security:
                - jwt_header_Authorization:
                    - api:access
    /user/{userID}/liketech:
        get:
            tags:
                - hy_usertech
            summary: getUserLikeTech hy_usertech
            description: get user's favorite techs
            operationId: hy_usertech#getUserLikeTech
            parameters:
                - name: userID
                  in: path
                  description: User ID
                  required: true
                  schema:
                    type: integer
                    description: User ID
                    example: 2328333941761142642
                    minimum: 1
                  example: 43820523876490080
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                description: 'Response body may alternatively be #/components/schemas/UsertechResponseTechCollection'
                                $ref: '#/components/schemas/UsertechCollection'
                            example:
                                - id: 10
                                  tech_name: Golang
                                - id: 10
                                  tech_name: Golang
            security:
                - jwt_header_Authorization:
                    - api:access
...
components:
    schemas:
...
        Usertech:
            type: object
            properties:
                id:
                    type: integer
                    description: ID
                    example: 10
                    minimum: 1
                tech_name:
                    type: string
                    description: Tech name
                    example: Golang
                    minLength: 1
                    maxLength: 40
            description: A user information
            example:
                id: 10
                tech_name: Golang
            required:
                - tech_name
        UsertechCollection:
            type: array
            items:
                $ref: '#/components/schemas/Usertech'
            example:
                - id: 10
                  tech_name: Golang
                - id: 10
                  tech_name: Golang
                - id: 10
                  tech_name: Golang
                - id: 10
                  tech_name: Golang
        UsertechResponseTech:
            type: object
            properties:
                tech_name:
                    type: string
                    description: Tech name
                    example: Golang
                    minLength: 1
                    maxLength: 40
            description: A user information (tech view)
            example:
                tech_name: Golang
            required:
                - tech_name
        UsertechResponseTechCollection:
            type: array
            items:
                $ref: '#/components/schemas/UsertechResponseTech'
            description: GetUserLikeTechResponseBody is the result type for an array of UsertechResponse (tech view)
            example:
                - tech_name: Golang
                - tech_name: Golang
...
    securitySchemes:
        jwt_header_Authorization:
            type: http
            scheme: bearer
...